refactor(database): extract conn.go (NewDBManager/getDB/initialize/detectDBEmbeddingDims) without behavior changes
